styling token box
