+++ ui/src/pages/Models/CDPToken.tsx
@@ -1,0 +1,123 @@
+/*
+ * CLOUDERA APPLIED MACHINE LEARNING PROTOTYPE (AMP)
+ * (C) Cloudera, Inc. 2025
+ * All rights reserved.
+ *
+ * Applicable Open Source License: Apache 2.0
+ *
+ * NOTE: Cloudera open source products are modular software products
+ * made up of hundreds of individual components, each of which was
+ * individually copyrighted.  Each Cloudera open source product is a
+ * collective work under U.S. Copyright Law. Your license to use the
+ * collective work is as provided in your written agreement with
+ * Cloudera.  Used apart from the collective work, this file is
+ * licensed for your use pursuant to the open source license
+ * identified above.
+ *
+ * This code is provided to you pursuant a written agreement with
+ * (i) Cloudera, Inc. or (ii) a third-party authorized to distribute
+ * this code. If you do not have a written agreement with Cloudera nor
+ * with an authorized and properly licensed third party, you do not
+ * have any rights to access nor to use this code.
+ *
+ * Absent a written agreement with Cloudera, Inc. ("Cloudera") to the
+ * contrary, A) CLOUDERA PROVIDES THIS CODE TO YOU WITHOUT WARRANTIES OF ANY
+ * KIND; (B) CLOUDERA DISCLAIMS ANY AND ALL EXPRESS AND IMPLIED
+ * WARRANTIES WITH RESPECT TO THIS CODE, INCLUDING BUT NOT LIMITED TO
+ * IMPLIED WARRANTIES OF TITLE, NON-INFRINGEMENT, MERCHANTABILITY AND
+ * FITNESS FOR A PARTICULAR PURPOSE; (C) CLOUDERA IS NOT LIABLE TO YOU,
+ * AND WILL NOT DEFEND, INDEMNIFY, NOR HOLD YOU HARMLESS FOR ANY CLAIMS
+ * ARISING FROM OR RELATED TO THE CODE; AND (D)WITH RESPECT TO YOUR EXERCISE
+ * OF ANY RIGHTS GRANTED TO YOU FOR THE CODE, CLOUDERA IS NOT LIABLE FOR ANY
+ * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, PUNITIVE OR
+ * CONSEQUENTIAL DAMAGES INCLUDING, BUT NOT LIMITED TO, DAMAGES
+ * RELATED TO LOST REVENUE, LOST PROFITS, LOSS OF INCOME, LOSS OF
+ * BUSINESS ADVANTAGE OR UNAVAILABILITY, OR LOSS OR CORRUPTION OF
+ * DATA.
+ */
+
+import { useState } from "react";
+import { useQueryClient } from "@tanstack/react-query";
+import { useSetCdpToken } from "src/api/ampMetadataApi.ts";
+import messageQueue from "src/utils/messageQueue.ts";
+import { QueryKeys } from "src/api/utils.ts";
+import { Button, Card, Flex, Input, Typography } from "antd";
+import { WarningOutlined } from "@ant-design/icons";
+
+export const CDPTokenInput = () => {
+  const [token, setToken] = useState("");
+  const queryClient = useQueryClient();
+  const setAuthToken = useSetCdpToken({
+    onSuccess: () => {
+      messageQueue.success("Token saved successfully");
+      queryClient
+        .invalidateQueries({
+          queryKey: [QueryKeys.getRerankingModels],
+        })
+        .catch(() => {
+          messageQueue.error("Error occurred fetching reranking models");
+        });
+      queryClient
+        .invalidateQueries({
+          queryKey: [QueryKeys.getLlmModels],
+        })
+        .catch(() => {
+          messageQueue.error("Error occurred fetching LLM models");
+        });
+      queryClient
+        .invalidateQueries({
+          queryKey: [QueryKeys.getEmbeddingModels],
+        })
+        .catch(() => {
+          messageQueue.error("Error occurred fetching embedding models");
+        });
+    },
+    onError: () => {
+      messageQueue.error("Error occurred setting token");
+    },
+  });
+
+  const handleSubmit = () => {
+    if (token) {
+      setAuthToken.mutate(token);
+    }
+  };
+
+  return (
+    <Card
+      title={
+        <Typography.Text type="danger">
+          <WarningOutlined style={{ marginRight: 8 }} />
+          Invalid or missing CDP token
+        </Typography.Text>
+      }
+    >
+      <Flex vertical gap={16}>
+        <Typography.Text italic>
+          Provide a valid CDP token to use Cloudera AI Inference
+        </Typography.Text>
+        <Flex gap={8}>
+          <Input
+            placeholder="CDP Token"
+            value={token}
+            onChange={(e) => {
+              setToken(e.target.value);
+            }}
+            onKeyDown={(e) => {
+              if (e.key === "Enter") {
+                handleSubmit();
+              }
+            }}
+          />
+          <Button
+            onClick={handleSubmit}
+            disabled={!token}
+            loading={setAuthToken.isPending}
+          >
+            Submit
+          </Button>
+        </Flex>
+      </Flex>
+    </Card>
+  );
+};
