+++ ui/src/pages/Models/ModelPage.tsx
@@ -36,10 +36,9 @@
  * DATA.
  ******************************************************************************/
 
-import { Alert, Button, Card, Flex, Input, Typography } from "antd";
+import { Flex, Typography } from "antd";
 import EmbeddingModelTable from "pages/Models/EmbeddingModelTable.tsx";
 import {
-  ModelSource,
   useGetEmbeddingModels,
   useGetLlmModels,
   useGetModelSource,
@@ -47,126 +46,12 @@ import {
 } from "src/api/modelsApi.ts";
 import InferenceModelTable from "pages/Models/InferenceModelTable.tsx";
 import RerankingModelTable from "pages/Models/RerankingModelTable.tsx";
-import { useState } from "react";
-import { useSetCdpToken } from "src/api/ampMetadataApi.ts";
-import messageQueue from "src/utils/messageQueue.ts";
-import { useQueryClient } from "@tanstack/react-query";
-import { ApiError, QueryKeys } from "src/api/utils.ts";
+import { CDPTokenInput } from "pages/Models/CDPToken.tsx";
+import {
+  checkHandledCaiiError,
+  ModelErrors,
+} from "pages/Models/ModelErrors.tsx";
 
-const ModelPageAlert = ({
-  error,
-  type,
-}: {
-  error: Error | null;
-  type: string;
-}) => {
-  if (!error) {
-    return null;
-  }
-  return (
-    <Alert
-      style={{ margin: 10 }}
-      message={`${type} model error: ${error.message}`}
-      type="error"
-    />
-  );
-};
-
-const CDPTokenInput = () => {
-  const [token, setToken] = useState("");
-  const queryClient = useQueryClient();
-  const setAuthToken = useSetCdpToken({
-    onSuccess: () => {
-      messageQueue.success("Token saved successfully");
-      queryClient
-        .invalidateQueries({
-          queryKey: [QueryKeys.getRerankingModels],
-        })
-        .catch(() => {
-          messageQueue.error("Error occurred fetching reranking models");
-        });
-      queryClient
-        .invalidateQueries({
-          queryKey: [QueryKeys.getLlmModels],
-        })
-        .catch(() => {
-          messageQueue.error("Error occurred fetching LLM models");
-        });
-      queryClient
-        .invalidateQueries({
-          queryKey: [QueryKeys.getEmbeddingModels],
-        })
-        .catch(() => {
-          messageQueue.error("Error occurred fetching embedding models");
-        });
-    },
-    onError: () => {
-      messageQueue.error("Error occurred setting token");
-    },
-  });
-
-  const handleSubmit = () => {
-    if (token) {
-      setAuthToken.mutate(token);
-    }
-  };
-
-  return (
-    <Card title="CDP Token Expired - Update to use Cloudera AI Inference Models">
-      <Flex gap={8}>
-        <Input
-          placeholder="CDP Token"
-          value={token}
-          onChange={(e) => {
-            setToken(e.target.value);
-          }}
-        />
-        <Button onClick={handleSubmit}>Submit</Button>
-      </Flex>
-    </Card>
-  );
-};
-
-const checkHandledCaiiError = (
-  inferenceError: Error | null | ApiError,
-  rerankingError: Error | null | ApiError,
-  embeddingError: Error | null | ApiError,
-) => {
-  return (
-    (inferenceError instanceof ApiError &&
-      (inferenceError.status === 401 || inferenceError.status === 500)) ||
-    (rerankingError instanceof ApiError &&
-      (rerankingError.status === 401 || rerankingError.status === 500)) ||
-    (embeddingError instanceof ApiError &&
-      (embeddingError.status === 401 || embeddingError.status === 500))
-  );
-};
-
-const ModelErrors = ({
-  inferenceError,
-  rerankingError,
-  embeddingError,
-  modelSource,
-}: {
-  inferenceError: Error | null;
-  embeddingError: Error | null;
-  rerankingError: Error | null;
-  modelSource?: ModelSource;
-}) => {
-  if (modelSource === "CAII") {
-    if (checkHandledCaiiError(inferenceError, rerankingError, embeddingError)) {
-      return null;
-    }
-  }
-
-  return (
-    <>
-      <ModelPageAlert error={inferenceError} type="Inference" />
-      <ModelPageAlert error={embeddingError} type="Embedding" />
-      <ModelPageAlert error={rerankingError} type="Reranking" />
-    </>
-  );
-};
 const ModelPage = () => {
   const {
     data: embeddingModels,
